% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_all_unique.R
\name{are_all_unique}
\alias{are_all_unique}
\title{Tests if all elements are unique}
\usage{
are_all_unique(x)
}
\arguments{
\item{x}{the elements to be tested}
}
\description{
Tests if all elements are unique
}
