% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  chunk_size,
  fasta_file,
  fasta_text,
  gene_ids,
  gene_name,
  hgvs,
  protein_id,
  protein_ids,
  snp_id,
  subject_ids,
  variation,
  variations,
  verbose
)
}
\arguments{
\item{chunk_size}{the number of protein IDs in an NCBI quey.
Too many values results in an NCBI API error}

\item{fasta_file}{name of a FASTA file}

\item{fasta_text}{text in the same format as a FASTA file}

\item{gene_ids}{one or more IDs of a gene}

\item{gene_name}{name of a gene}

\item{hgvs}{a variation in HGVS ('Human Genome Variation Society') format}

\item{protein_id}{ID of a protein.
For example, \code{NP_001007554.1} is a protein ID.}

\item{protein_ids}{one or more protein IDs.
For example, \code{NP_001007554.1} is a protein ID.}

\item{snp_id}{the ID of a SNP}

\item{subject_ids}{one of more IDs of a subject}

\item{variation}{one variation.
For example, \code{NP_009089.4:p.Val723Gly} is a variation.}

\item{variations}{one or more variations.
For example, \code{NP_009089.4:p.Val723Gly} is a variation.}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
\code{@noRd}. This is not done, as this will disallow all
functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
