% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_n_sequences_from_unique_protein_ids.R
\name{fetch_n_sequences_from_unique_protein_ids}
\alias{fetch_n_sequences_from_unique_protein_ids}
\title{Internal function}
\usage{
fetch_n_sequences_from_unique_protein_ids(
  protein_ids,
  max_n_protein_ids = 100,
  verbose = FALSE
)
}
\arguments{
\item{protein_ids}{one or more protein IDs.
For example, \code{NP_001007554.1} is a protein ID.}

\item{max_n_protein_ids}{the maximum number of protein IDs}

\item{verbose}{set to TRUE for more output}
}
\value{
a character vector with protein sequences,
one per protein ID
}
\description{
Fetch the n sequence from one or more unique protein ID,
where \code{max_n_protein_ids}
is the maximum number of sequences NCBI allows for
}
