% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequence_from_protein_id.R
\name{fetch_sequence_from_protein_id}
\alias{fetch_sequence_from_protein_id}
\title{Fetch the sequence from a protein ID}
\usage{
fetch_sequence_from_protein_id(protein_id, verbose = FALSE)
}
\arguments{
\item{protein_id}{ID of a protein.
For example, \code{NP_001007554.1} is a protein ID.}

\item{verbose}{set to TRUE for more output}
}
\description{
Fetch the protein sequence using a protein ID
}
