% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequences_from_unique_protein_ids.R
\name{fetch_sequences_from_unique_protein_ids}
\alias{fetch_sequences_from_unique_protein_ids}
\title{Fetch the sequence from one or more protein ID}
\usage{
fetch_sequences_from_unique_protein_ids(
  protein_ids,
  chunk_size = 100,
  verbose = FALSE
)
}
\arguments{
\item{protein_ids}{one or more protein IDs.
For example, \code{NP_001007554.1} is a protein ID.}

\item{chunk_size}{the number of protein IDs in an NCBI quey.
Too many values results in an NCBI API error}

\item{verbose}{set to TRUE for more output}
}
\value{
a character vector with protein sequences,
one per protein ID
}
\description{
Fetch the protein sequence using one or more protein IDs
}
