% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_names_from_human_gene_ids.R
\name{get_gene_names_from_human_gene_ids}
\alias{get_gene_names_from_human_gene_ids}
\title{From a human gene ID, get the gene name.
This function checks if the gene is indeed fror a human}
\usage{
get_gene_names_from_human_gene_ids(gene_ids, verbose = FALSE)
}
\arguments{
\item{gene_ids}{one or more IDs of a gene}

\item{verbose}{set to TRUE for more output}
}
\description{
From a human gene ID, get the gene name.
This function checks if the gene is indeed fror a human
}
